WIP: Prevent accessing array and hash parameters via square brackets
Today, when accessing parameters using `params[:key]` you can recieve either a single value like a string or a number, or you might recieve an array or hash. This can result in confusing behavior when you access a key directly and it is not the type you expect.

This PR changes the behavior of square brackets on params to act more like calling `params.permit(:key)[:key]` where hashes and arrays are replaced with nil since they where not explicitly allowed via strong params.

This is a breaking change, so my suggestion is this be introduced behind a configuration option in rails 7.0 and default to on in new rails apps